refactor: move servicoscontent to its own directory
