+++ src/screens/public/PartnerProfile/ServicosContent/index.ts
@@ -1,0 +1,1 @@
+export { ServicosContent } from './ServicosContent';
